# Conflicts:
#	DM/DM.DAL/bin/Debug/net5.0/DM.DAL.dll
#	DM/DM.DAL/bin/Debug/net5.0/DM.DAL.pdb
#	DM/DM.DAL/obj/DM.DAL.csproj.nuget.dgspec.json
#	DM/DM.DAL/obj/Debug/net5.0/DM.DAL.GeneratedMSBuildEditorConfig.editorconfig
#	DM/DM.DAL/obj/Debug/net5.0/DM.DAL.assets.cache
#	DM/DM.DAL/obj/Debug/net5.0/DM.DAL.csproj.CoreCompileInputs.cache
#	DM/DM.DAL/obj/Debug/net5.0/DM.DAL.dll
#	DM/DM.DAL/obj/Debug/net5.0/DM.DAL.genruntimeconfig.cache
#	DM/DM.DAL/obj/Debug/net5.0/DM.DAL.pdb
#	DM/DM.DAL/obj/project.assets.json
#	DM/DM.DAL/obj/project.nuget.cache
#	DM/DM.DAL/obj/project.packagespec.json
#	DM/DM.DAL/obj/rider.project.restore.info
#	DM/DM.Domain/bin/Debug/net5.0/DM.DAL.dll
#	DM/DM.Domain/bin/Debug/net5.0/DM.DAL.pdb
#	DM/DM.Domain/bin/Debug/net5.0/DM.Domain.dll
#	DM/DM.Domain/bin/Debug/net5.0/DM.Domain.pdb
#	DM/DM.Domain/obj/DM.Domain.csproj.nuget.dgspec.json
#	DM/DM.Domain/obj/Debug/net5.0/DM.Domain.GeneratedMSBuildEditorConfig.editorconfig
#	DM/DM.Domain/obj/Debug/net5.0/DM.Domain.assets.cache
#	DM/DM.Domain/obj/Debug/net5.0/DM.Domain.csproj.AssemblyReference.cache
#	DM/DM.Domain/obj/Debug/net5.0/DM.Domain.csproj.CoreCompileInputs.cache
#	DM/DM.Domain/obj/Debug/net5.0/DM.Domain.dll
#	DM/DM.Domain/obj/Debug/net5.0/DM.Domain.pdb
#	DM/DM.Domain/obj/project.assets.json
#	DM/DM.Domain/obj/project.nuget.cache
#	DM/DM.Domain/obj/project.packagespec.json
#	DM/DM.Domain/obj/rider.project.restore.info
#	DM/DM/bin/Debug/net5.0/DM.DAL.dll
#	DM/DM/bin/Debug/net5.0/DM.DAL.pdb
#	DM/DM/bin/Debug/net5.0/DM.Domain.dll
#	DM/DM/bin/Debug/net5.0/DM.Domain.pdb
#	DM/DM/bin/Debug/net5.0/DM.dll
#	DM/DM/bin/Debug/net5.0/DM.pdb
#	DM/DM/obj/DM.csproj.nuget.dgspec.json
#	DM/DM/obj/Debug/net5.0/DM.GeneratedMSBuildEditorConfig.editorconfig
#	DM/DM/obj/Debug/net5.0/DM.assets.cache
#	DM/DM/obj/Debug/net5.0/DM.csproj.AssemblyReference.cache
#	DM/DM/obj/Debug/net5.0/DM.csproj.CoreCompileInputs.cache
#	DM/DM/obj/Debug/net5.0/DM.csproj.FileListAbsolute.txt
#	DM/DM/obj/Debug/net5.0/DM.dll
#	DM/DM/obj/Debug/net5.0/DM.genruntimeconfig.cache
#	DM/DM/obj/Debug/net5.0/DM.pdb
#	DM/DM/obj/Debug/net5.0/ref/DM.dll
#	DM/DM/obj/Debug/net5.0/refint/DM.dll
#	DM/DM/obj/project.assets.json
#	DM/DM/obj/project.nuget.cache
#	DM/DM/obj/project.packagespec.json
#	DM/DM/obj/rider.project.restore.info

+++ DM/DM.DAL/Entities/UserProjectEntity.cs
@@ -7,11 +7,11 @@ namespace DM.DAL.Entities
     /// </summary>
     public class UserProjectEntity : BaseEntity
     {
-        public int UserId { get; set; }
+        public long UserId { get; set; }
 
         public UserEntity User { get; set; }
 
-        public int ProjectId { get; set; }
+        public long ProjectId { get; set; }
 
         public ProjectEntity Project { get; set; }
     }
